moved some from wmexchanger suggested by Keith
