+++ eidoscommon/src/main/scala/org/clulab/wm/eidoscommon/utils/ThreadUtils.scala
@@ -27,4 +27,17 @@ object ThreadUtils {
     parSeq.tasksupport = forkJoinTaskSupport
     parSeq
   }
+
+  def stop(thread: Thread, duration: Long): Unit = {
+    try {
+      thread.interrupt
+      thread.join(duration)
+      if (thread.isAlive)
+        thread.stop
+    }
+    catch {
+      case _: InterruptedException =>
+      case _: ThreadDeath =>
+    }
+  }
 }
