--- wmexchanger/src/main/scala/org/clulab/wm/wmexchanger/utils/ThreadUtils.scala
@@ -1,17 +1,0 @@
-package org.clulab.wm.wmexchanger.utils
-
-object ThreadUtils {
-
-  def stop(thread: Thread, duration: Long): Unit = {
-    try {
-      thread.interrupt
-      thread.join(duration)
-      if (thread.isAlive)
-        thread.stop
-    }
-    catch {
-      case _: InterruptedException =>
-      case _: ThreadDeath =>
-    }
-  }
-}
