+++ wmexchanger/src/main/scala/org/clulab/wm/wmexchanger/wmconsumer/RestConsumerApp.scala
@@ -13,13 +13,8 @@ import org.apache.http.client.utils.URIBuilder
 import org.apache.http.impl.client.BasicCredentialsProvider
 import org.apache.http.impl.client.CloseableHttpClient
 import org.apache.http.impl.client.HttpClientBuilder
-import org.clulab.wm.wmexchanger.utils.Closer.AutoCloser
-import org.clulab.wm.wmexchanger.utils.FileUtils
-import org.clulab.wm.wmexchanger.utils.FileEditor
-import org.clulab.wm.wmexchanger.utils.PropertiesBuilder
-import org.clulab.wm.wmexchanger.utils.Sinker
-import org.clulab.wm.wmexchanger.utils.Sourcer
-import org.clulab.wm.wmexchanger.utils.StringUtils
+import org.clulab.wm.eidoscommon.utils.{FileEditor, FileUtils, PropertiesBuilder, Sinker, Sourcer, StringUtils}
+import org.clulab.wm.eidoscommon.utils.Closer.AutoCloser
 import org.json4s.DefaultFormats
 import org.json4s.jackson.JsonMethods
 import org.slf4j.Logger
